refactor(agents): pin OpenInferenceAgentWrapper to ChatDependencies/ChatOutput
Drop the unused AgentDepsT/OutputDataT generics from the wrapper and bind
it directly to the only types it's ever instantiated with. Move ChatOutput
next to ChatDependencies so both the wrapper and agent_factory can import
it without a circular dependency.
