+++ src/phoenix/server/agents/dependencies.py
@@ -1,5 +1,7 @@
 from dataclasses import dataclass
+from typing import TypeAlias
 
+from pydantic_ai import DeferredToolRequests
 from pydantic_ai.mcp import MCPServerStreamableHTTP
 
 from phoenix.server.agents.capabilities import AgentCapabilities
@@ -11,3 +13,6 @@ class ChatDependencies:
     contexts: ResolvedContexts
     capabilities: AgentCapabilities
     docs_mcp_toolset: MCPServerStreamableHTTP | None
+
+
+ChatOutput: TypeAlias = str | DeferredToolRequests
